refactor(bedrock): extract inline tests to sibling service_tests.rs
service.rs (2908 lines) was mostly inline #[cfg(test)] mod tests {} at
1672-2908. Move that 1237-line test body to service_tests.rs and
reference via #[path].

service.rs: 2908 -> 1676 lines, under 2000-line audit threshold.
332 lib tests still pass.
